Refactor: Move enums out of SeaORM generated directory
