+++ deepwell/src/models/file_revision.rs
@@ -1,6 +1,6 @@
 //! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.19
 
-use super::sea_orm_active_enums::FileRevisionType;
+use crate::types::FileRevisionType;
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
