--- deepwell/src/models/sea_orm_active_enums.rs
@@ -1,259 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.19
-
-use sea_orm::{TryFromU64, entity::prelude::*};
-use serde::{Deserialize, Serialize};
-use strum_macros::{Display, EnumString};
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum AliasType {
-    Site,
-    User,
-}
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum FileRevisionType {
-    Regular,
-    Rollback,
-    Create,
-    Delete,
-    Undelete,
-    Move,
-}
-#[derive(
-    Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum, Copy, Serialize, Deserialize,
-)]
-#[sea_orm(rs_type = "String", db_type = "String(StringLen::None)")]
-#[serde(rename_all = "kebab-case")]
-pub enum License {
-    #[sea_orm(string_value = "cc-by-sa-4.0")]
-    #[serde(rename = "cc-by-sa-4.0")]
-    CcBySa40,
-    #[sea_orm(string_value = "cc-by-4.0")]
-    #[serde(rename = "cc-by-4.0")]
-    CcBy40,
-    #[sea_orm(string_value = "cc-by-nd-4.0")]
-    #[serde(rename = "cc-by-nd-4.0")]
-    CcByNd40,
-    #[sea_orm(string_value = "cc-by-nc-4.0")]
-    #[serde(rename = "cc-by-nc-4.0")]
-    CcByNc40,
-    #[sea_orm(string_value = "cc-by-nc-sa-4.0")]
-    #[serde(rename = "cc-by-nc-sa-4.0")]
-    CcByNcSa40,
-    #[sea_orm(string_value = "cc-by-nc-nd-4.0")]
-    #[serde(rename = "cc-by-nc-nd-4.0")]
-    CcByNcNd40,
-    #[sea_orm(string_value = "cc-by-sa-3.0")]
-    #[serde(rename = "cc-by-sa-3.0")]
-    CcBySa30,
-    #[sea_orm(string_value = "cc-by-3.0")]
-    #[serde(rename = "cc-by-3.0")]
-    CcBy30,
-    #[sea_orm(string_value = "cc-by-nd-3.0")]
-    #[serde(rename = "cc-by-nd-3.0")]
-    CcByNd30,
-    #[sea_orm(string_value = "cc-by-nc-3.0")]
-    #[serde(rename = "cc-by-nc-3.0")]
-    CcByNc30,
-    #[sea_orm(string_value = "cc-by-nc-sa-3.0")]
-    #[serde(rename = "cc-by-nc-sa-3.0")]
-    CcByNcSa30,
-    #[sea_orm(string_value = "cc-by-nc-nd-3.0")]
-    #[serde(rename = "cc-by-nc-nd-3.0")]
-    CcByNcNd30,
-    #[sea_orm(string_value = "cc-by-sa-2.5")]
-    #[serde(rename = "cc-by-sa-2.5")]
-    CcBySa25,
-    #[sea_orm(string_value = "cc-by-2.5")]
-    #[serde(rename = "cc-by-2.5")]
-    CcBy25,
-    #[sea_orm(string_value = "cc-by-nd-2.5")]
-    #[serde(rename = "cc-by-nd-2.5")]
-    CcByNd25,
-    #[sea_orm(string_value = "cc-by-nc-2.5")]
-    #[serde(rename = "cc-by-nc-2.5")]
-    CcByNc25,
-    #[sea_orm(string_value = "cc-by-nc-sa-2.5")]
-    #[serde(rename = "cc-by-nc-sa-2.5")]
-    CcByNcSa25,
-    #[sea_orm(string_value = "cc-by-nc-nd-2.5")]
-    #[serde(rename = "cc-by-nc-nd-2.5")]
-    CcByNcNd25,
-    #[sea_orm(string_value = "gnu-fdl-1.3")]
-    #[serde(rename = "gnu-fdl-1.3")]
-    GnuFdl13,
-    #[sea_orm(string_value = "gnu-fdl-1.2")]
-    #[serde(rename = "gnu-fdl-1.2")]
-    GnuFdl12,
-    #[sea_orm(string_value = "gnu-fdl-1.1")]
-    #[serde(rename = "gnu-fdl-1.1")]
-    GnuFdl11,
-    #[sea_orm(string_value = "cc0")]
-    #[serde(rename = "cc0")]
-    Cc0,
-}
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum MessageRecipientType {
-    Regular,
-    Cc,
-    Bcc,
-}
-
-impl TryFromU64 for MessageRecipientType {
-    fn try_from_u64(_: u64) -> std::result::Result<Self, DbErr> {
-        Err(DbErr::ConvertFromU64(
-            "cannot construct MessageRecipientType from u64; auto_increment must be false for this primary key",
-        ))
-    }
-}
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum PageRevisionType {
-    Regular,
-    Rollback,
-    Undo,
-    Create,
-    Delete,
-    Undelete,
-    Move,
-}
-
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum RelationObjectType {
-    Site,
-    User,
-    Page,
-    File,
-}
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum TextBlockType {
-    Code,
-    Html,
-}
-impl TryFromU64 for TextBlockType {
-    fn try_from_u64(_: u64) -> std::result::Result<Self, DbErr> {
-        Err(DbErr::ConvertFromU64(
-            "cannot construct TextBlockType from u64; auto_increment must be false for this primary key",
-        ))
-    }
-}
-#[derive(
-    EnumIter,
-    Serialize,
-    Deserialize,
-    Debug,
-    Copy,
-    Clone,
-    Hash,
-    PartialEq,
-    Eq,
-    DeriveValueType,
-    EnumString,
-    Display,
-)]
-#[sea_orm(value_type = "String")]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab_case", ascii_case_insensitive)]
-pub enum UserType {
-    Regular,
-    System,
-    Site,
-    Bot,
-}
