+++ deepwell/src/models/text_block.rs
@@ -1,6 +1,6 @@
 //! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.19
 
-use super::sea_orm_active_enums::TextBlockType;
+use crate::types::TextBlockType;
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
