+++ deepwell/src/services/alias/service.rs
@@ -20,11 +20,11 @@
 
 use super::prelude::*;
 use crate::models::alias::{self, Entity as Alias, Model as AliasModel};
-use crate::models::sea_orm_active_enums::AliasType;
 use crate::models::site::{self, Entity as Site};
 use crate::models::user::{self, Entity as User};
 use crate::services::filter::{FilterClass, FilterType};
 use crate::services::{FilterService, SiteService, UserService};
+use crate::types::AliasType;
 use crate::types::Reference;
 use crate::utils::get_regular_slug;
