+++ deepwell/src/services/alias/structs.rs
@@ -18,7 +18,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-use crate::models::sea_orm_active_enums::AliasType;
+use crate::types::AliasType;
 
 #[derive(Deserialize, Debug)]
 pub struct CreateAlias {
