+++ deepwell/src/services/caddy/service.rs
@@ -25,11 +25,11 @@
 
 use super::prelude::*;
 use crate::models::alias::Model as AliasModel;
-use crate::models::sea_orm_active_enums::AliasType;
 use crate::models::site::{self, Entity as Site};
 use crate::models::site_domain::{self, Entity as SiteDomain};
 use crate::services::domain::DEFAULT_SITE_SLUG;
 use crate::services::{AliasService, DomainService};
+use crate::types::AliasType;
 use askama::Template;
 use sea_orm::{EntityTrait, QuerySelect};
 use std::borrow::Cow;
