+++ deepwell/src/services/file_revision/mod.rs
@@ -23,7 +23,7 @@ mod prelude {
     pub use super::super::prelude::*;
     pub use super::structs::*;
     pub use crate::hash::BlobHash;
-    pub use crate::models::sea_orm_active_enums::FileRevisionType;
+    pub use crate::types::FileRevisionType;
 }
 
 mod service;
