+++ deepwell/src/services/relation/site_user.rs
@@ -28,8 +28,8 @@
 //! As such, it is an invariant that all users linked here are of the type `site`.
 
 use super::prelude::*;
-use crate::models::sea_orm_active_enums::UserType;
 use crate::services::UserService;
+use crate::types::UserType;
 
 impl_relation!(SiteUser, Site, site_id, User, user_id, (), NO_CREATE_IMPL);
